+++ src/cfnlint/rules/outputs/Description.py
@@ -1,0 +1,44 @@
+"""
+  Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+  Permission is hereby granted, free of charge, to any person obtaining a copy of this
+  software and associated documentation files (the "Software"), to deal in the Software
+  without restriction, including without limitation the rights to use, copy, modify,
+  merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
+  permit persons to whom the Software is furnished to do so.
+
+  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+  PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+  OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+  SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+"""
+import six
+from cfnlint import CloudFormationLintRule
+from cfnlint import RuleMatch
+
+
+class Description(CloudFormationLintRule):
+    """Check if Outputs Descriptions are only string values"""
+    id = 'E6005'
+    shortdesc = 'Outputs descriptions can only be strings'
+    description = 'Outputs descriptions can only be strings'
+    source_url = 'https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/outputs-section-structure.html'
+    tags = ['outputs']
+
+    def match(self, cfn):
+        """Check CloudFormation Outputs"""
+
+        matches = list()
+
+        outputs = cfn.template.get('Outputs', {})
+        if outputs:
+            for output_name, output_value in outputs.items():
+                description = output_value.get('Description')
+                if description:
+                    if not isinstance(description, six.string_types):
+                        message = 'Output Description can only be a string'
+                        matches.append(RuleMatch(['Outputs', output_name, 'Description'], message))
+
+        return matches